Move password component to connected directory
